+++ lang/src/org/partiql/lang/planner/PlannerPipeline.kt
@@ -23,10 +23,12 @@ import org.partiql.lang.errors.Problem
 import org.partiql.lang.errors.ProblemCollector
 import org.partiql.lang.errors.ProblemHandler
 import org.partiql.lang.errors.Property
+import org.partiql.lang.eval.CompileOptions
 import org.partiql.lang.eval.ExprFunction
 import org.partiql.lang.eval.ExprValueFactory
 import org.partiql.lang.eval.Expression
 import org.partiql.lang.eval.ThunkReturnTypeAssertions
+import org.partiql.lang.eval.TypedOpBehavior
 import org.partiql.lang.eval.builtins.DynamicLookupExprFunction
 import org.partiql.lang.eval.builtins.createBuiltinFunctions
 import org.partiql.lang.eval.builtins.storedprocedure.StoredProcedure
@@ -45,6 +47,7 @@ import org.partiql.lang.eval.physical.operators.RelationalOperatorFactoryKey
 import org.partiql.lang.eval.physical.operators.ScanRelationalOperatorFactoryDefault
 import org.partiql.lang.eval.physical.operators.SortOperatorFactoryDefault
 import org.partiql.lang.eval.physical.operators.UnpivotOperatorFactoryDefault
+import org.partiql.lang.eval.visitors.PartiqlAstSanityValidator
 import org.partiql.lang.planner.transforms.normalize
 import org.partiql.lang.planner.transforms.toDefaultPhysicalPlan
 import org.partiql.lang.planner.transforms.toLogicalPlan
@@ -511,6 +514,7 @@ internal class PlannerPipelineImpl(
         val normalizedAst = plannerEventCallback.doEvent("normalize_ast", ast) {
             ast.normalize()
         }
+        normalizedAst.validate(evaluatorOptions.typedOpBehavior)
 
         // ast -> logical plan
         val logicalPlan = plannerEventCallback.doEvent("ast_to_logical", normalizedAst) {
@@ -572,6 +576,15 @@ internal class PlannerPipelineImpl(
             }
         // If we reach this far, we're successful.  If there were any problems at all, they were just warnings.
         return PlannerPassResult.Success(finalPlan, problemHandler.problems)
+    }
+
+    /**
+     * Performs a validation of the AST. The [PartiqlAstSanityValidator] only requires the
+     * [TypedOpBehavior] to perform assertions, so we pass this along.
+     */
+    private fun PartiqlAst.Statement.validate(behavior: TypedOpBehavior) {
+        val validatorCompileOptions = CompileOptions.build { typedOpBehavior(behavior) }
+        PartiqlAstSanityValidator().validate(this, validatorCompileOptions)
     }
 
     override fun compile(physicalPlan: PartiqlPhysical.Plan): PlannerPassResult<QueryPlan> =